prinft

+++ main.h
@@ -8,6 +8,6 @@
 int _printf(const char *format, ...);
 int _putchar(char c);
 int print_str(char *c);
-int print_int(long n);
-int print_i_int(long n);
+long print_int(long n);
+long print_i_int(long n);
 #endif
